[Credits] Focus Back button on open, allow scrolling with ui_(page)up/down with tweening
Note that we had to make the Rich Text Label Fit Content so the parent ScrollContainer could recognize its size and actually show its scrollbar, so we can set it by code. The scrollbar we saw so far was the integrated scrollbar of the text, which was unrelated. For safety we also disabled Scroll Active, but the text's own scrollbar won't show as long as the text is contained in some ScrollContainer anyway.