[Pause] Fix Options Menu receiving input and preventing Pause Menu to be open on first ui_cancel press
and also showing with default values of 0, then saving them on Back causing Volume to be muted and fullscreen/Vsync config to be lost.

This happened only when OptionsMenu was visible in scene, but scene we sometimes do it for testing, we prefer more robust code to handle this case.